requirements.txt
# Core
pydantic>=2.0.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
fastapi>=0.100.0
uvicorn>=0.23.0

# Network/Data
requests>=2.31.0
arxiv>=2.1.0
google-generativeai>=0.3.0
openai>=1.0.0

# Dev/Test
pytest>=7.0.0
ruff>=0.1.0
